% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_Par.R
\name{AF_Par}
\alias{AF_Par}
\title{Alkire-Foster (AF) method in parallel mode}
\usage{
AF_Par(df, g = NULL, w = NULL, k = 1)
}
\arguments{
\item{df}{A poverty data frame}

\item{g}{A groups(default as NULL)}

\item{w}{An indicator weight vectors (default as 1)}

\item{k}{A poverty cut-off (default as 1)}
}
\value{
returns a \code{list} of \code{list} contains

\item{H}{Uncensored head count ratio is the proportion of the population that is multidimensionally deprived}
\item{A}{Average deprivation share among poors}
\item{M0}{Multidimensional Poverty Index}
\item{DimentionalContribution}{Dimensional contributions data frame}
\item{pov_df}{poverty data frame associated with each group}
}
\description{
Multidimensional Poverty Index using Alkire-Foster (AF) method
which running in parallel mode
}
