% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_Seq.R
\name{AF_Seq}
\alias{AF_Seq}
\title{Alkire-Foster (AF) method in sequential mode}
\usage{
AF_Seq(df, g = NULL, w = NULL, k = 1)
}
\arguments{
\item{df}{A poverty data frame}

\item{g}{A groups(default as NULL)}

\item{w}{An indicator weight vectors (default as 1)}

\item{k}{A poverty cut-off (default as 1)}
}
\value{
returns a \code{list} of \code{list} contains

\item{H}{Uncensored head count ratio the proportion of the population that is multidimensionally deprived calculated by divide the number of poor people with the total number of people.}
\item{A}{Average deprivation share among poor people, by aggregating the proportion of total deprivations each person and dividing by the total number of poor people.}
\item{M0}{Multidimensional Poverty Index (MPI)}
\item{DimentionalContribution}{Dimensional contributions denotes the magnitude of each indicators impact on MPI.}
\item{pov_df}{poverty data frame associated with each group.\code{Cvector} is total deprived indicators. \code{IsPoverty} is a binary variable with only 1 and 0, with 1 indicating that person does not meet the threshold(poor person) and 0 indicating the opposite. \code{Intensity}, The intensity of a deprived indication among impoverished people is computed by dividing the number of deprived indicators by the total number of indicators.}
}
\description{
Multidimensional Poverty Index using Alkire-Foster (AF) method
computes in sequential mode
}
