% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{examplePovertydf}
\alias{examplePovertydf}
\title{A simulation poverty data frame}
\format{
A data frame with 30 rows of individuals and 17 variables:
\describe{
    The first 16 variables are deprivation indicators where any indicators is equal to 1 means deprived and 0 means not deprived.
    \itemize{
        \item{Access.to.health.care}{}
        \item{Child.Mortality}{}
        \item{Access.to.clean.source.of.water}{}
        \item{Access.to.an.improve.sanatation}{}
        \item{Electricity}{}
        \item{Asset.ownership}{}
        \item{Cooking.Fuel}{}
        \item{Housing.Materials}{}
        \item{School.lag}{}
        \item{School.attendance}{}
        \item{Years.of.education}{}
    }
    The last variable is simulated region from France (forth-level administrative division).
    \itemize{
        \item{Region}{}
    }
}
}
\usage{
examplePovertydf
}
\description{
A simulation poverty data frame
}
\keyword{datasets}
