% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.code.R
\name{state.code}
\alias{state.code}
\title{A function to fill in iso3c codes for Mexico's states.}
\usage{
state.code(x, drop.state = T)
}
\arguments{
\item{x}{dataframe with a column of state names called \code{state}}

\item{drop.state}{Drops the state column after inserting the `geocodes` column. \code{TRUE} by default.}
}
\description{
A function to fill in iso3c codes for Mexico's states.
}
