% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI.R
\name{MPI}
\alias{MPI}
\title{Multidimensional Poverty Index}
\usage{
MPI(households, data)
}
\arguments{
\item{households}{a vector (1 x n)}

\item{data}{a vector (1 x n) n={10,20,30, ...}}
}
\value{
H, A and MPI
}
\description{
Multidimensional Poverty Index
}
